Fixed radix join's handling of partitions and delta
Previously, the radix join would assume that the table is a row store.
It would use the offset of a column in a table as the offset in the
attribute vector. That will fail for most other situations, especially
with a pure column store.

Furthermore, the handling of delta was fixed. The old cold did not
retrieve the actual attribute vector of the delta but the main instead.